refactor :MentoArticlePanel 영역 리팩토링했습니다
